zenith_bot (1)/zenith_bot/requirements.txt
# Zenith Bot — Requirements
# Python 3.11+

# Telegram Bot Framework
aiogram==3.7.0

# Database
asyncpg==0.29.0
supabase==2.4.0

# AI Providers
openai==1.30.0
groq==0.9.0
google-generativeai==0.7.0
httpx==0.27.0
aiohttp==3.9.5

# Data
tvdatafeed==2.3.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
ta==0.11.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Web (Google OAuth callback)
fastapi==0.111.0
uvicorn==0.30.1

# RAG
cohere==5.5.0

# News
feedparser==6.0.11
gnews==0.3.7

# Payment
midtransclient==1.4.2

# Scheduler
apscheduler==3.10.4

# Image generation (chart)
matplotlib==3.9.0
mplfinance==0.12.10b0
Pillow==10.3.0

# Utils
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.7.3
pydantic-settings==2.3.0
tenacity==8.3.0
pytz==2024.1
